rename svg-components.ts to svg-general-components.ts to reflect it's not the onlu components file
