+++ src/ts/waterfall/svg-general-components.ts
